fix(Identity): removed unused method
