+++ docs/platform/identities.md
@@ -10,23 +10,18 @@ Decoupling from the Blockchain identity allow to provide a unique user experienc
 Identity can be of multiple type, supported by this SDK, you will find these two types : 
    
 - User Identity as type `user`
-
 - Application Identity as type `application`
 
 
 ```js
-const bob = sdk.platform.identities.register('user', {});
+const alice = sdk.platform.identities.register();
+const bob = sdk.platform.identities.register('user');
+const appId = sdk.platform.identities.register('application');
 ```   
 
-
-## Searching an identity
-
-```js
-const bob = sdk.platform.identities.search('user', {});
-```   
 
 ## Get an identity
 
 ```js
-const bob = sdk.platform.identities.get('user', 'bob');
+const bob = sdk.platform.identities.get('3GegupTgRfdN9JMS8R6QXF3B2VbZtiw63eyudh1oMJAk');
 ```
