--- src/DashJS/SDK/Platform/methods/identities/search.ts
@@ -1,8 +1,0 @@
-import {Platform} from "../../Platform";
-
-
-export function search(this: Platform): any {
-    throw new Error('Implementation missing in dependencies.');
-}
-
-export default search;
